Update delete.sh delete2.sh delete3.sh again

+++ bash/delete2.sh
@@ -4,25 +4,25 @@
 ### This is configuration for .vimrc file
 
 ## Delete .vimrc and .bashrc in root folder 
-
 if [ -f ~/.vimrc ]
 then
 	sudo rm -f ~/.vimrc
 fi
 
-if [ -f ~/.bashrc ]
+if [ -f ~/.bashrc]
 then
 	sudo rm -f ~/.bashrc
 fi
 
 ## If .vimrc in root folder does not exist. mv it from preconfig/backup folder
-if [ ! -f ~/.vimrc ]
+if [ ! -f ~/.vimrc ] && [ -f ~/preconfig/backup/.vimrc.backup ]
 then
 	mv ~/preconfig/backup/.vimrc.backup ~/.vimrc
 fi
 
 ## If .bashrc in root folder does not exist. mv it from preconfig/backup folder
-if [ ! -f ~/.bashrc ]
+if [ ! -f ~/.bashrc ] && [ -f ~/preconfig/backup/.bashrc.backup ]
+
 then
 	mv ~/preconfig/backup/.bashrc.backup ~/.bashrc
 fi
@@ -42,7 +42,4 @@ if [ -d ~/ArgoCD ]
 then
 	rm -rf ~/AgoCD
 fi
-
-## Clear all alias
-exec bash
